Refactored data pushes to CCOS to use a single script.
