--- .github/workflows/generate_projects_page.yml
@@ -1,28 +1,0 @@
-name: Generate Projects Page
-on: 
-  schedule:
-  - cron: '0 0 * * *'
-  push:
-    branches:
-      - master
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-    - name: Set up Python 3.7
-      uses: actions/setup-python@v1
-      with:
-        python-version: 3.7
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        python -m pip install pipenv
-        cd generate_projects_page
-        pipenv install
-    - name: Export token to env and run the script
-      run: |
-        cd generate_projects_page
-        pipenv run python generate_projects_page.py
-      env:
-        ADMIN_GITHUB_TOKEN: ${{ secrets.ADMIN_GITHUB_TOKEN }}
