--- generate_projects_page/Pipfile
@@ -1,35 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-astroid = "==2.3.3"
-certifi = "==2019.11.28"
-cffi = "==1.14.0"
-chardet = "==3.0.4"
-cryptography = "==2.8"
-emoji = "==0.5.4"
-gitdb = "==4.0.2"
-idna = "==2.9"
-isort = "==4.3.21"
-lazy-object-proxy = "==1.4.3"
-mccabe = "==0.6.1"
-pycparser = "==2.20"
-pygithub = "==1.46"
-pylint = "==2.4.4"
-requests = "==2.23.0"
-six = "==1.14.0"
-smmap = "==3.0.1"
-typed-ast = "==1.4.1"
-urllib3 = "==1.25.8"
-wrapt = "==1.11.2"
-Deprecated = "==1.2.7"
-GitPython = "==3.1.0"
-PyJWT = "==1.7.1"
-PyYAML = "==5.3"
-
-[requires]
-python_version = "3.7"
