+++ push_data_to_ccos/get_repo_data.py
@@ -1,58 +1,17 @@
 #!/usr/bin/env python3
 # vim: set fileencoding=utf-8:
-# Standard library
-import json
-import os
 
 # Third-party
 from github import Github
-from github.GithubException import UnknownObjectException
+from github.GithubException import GithubException, UnknownObjectException
 import emoji
-import git
 import yaml
 
-
-GIT_USER_NAME = "CC creativecommons.github.io Bot"
-GIT_USER_EMAIL = "cc-creativecommons-github-io-bot@creativecommons.org"
-
-GITHUB_USERNAME = "cc-creativecommons-github-io-bot"
-GITHUB_ORGANIZATION = "creativecommons"
-GITHUB_REPO_NAME = "creativecommons.github.io-source"
+# Local
+from push_data_via_git import GITHUB_ORGANIZATION, GITHUB_TOKEN
 
 
-GITHUB_TOKEN = os.environ["ADMIN_GITHUB_TOKEN"]
-GITHUB_REPO_URL_WITH_CREDENTIALS = (
-    f"https://{GITHUB_USERNAME}:{GITHUB_TOKEN}"
-    f"@github.com/{GITHUB_ORGANIZATION}/{GITHUB_REPO_NAME}.git"
-)
 CC_METADATA_FILE_NAME = ".cc-metadata.yml"
-
-WORKING_DIRECTORY = "/tmp"
-GIT_WORKING_DIRECTORY = f"{WORKING_DIRECTORY}/{GITHUB_REPO_NAME}"
-JSON_FILE_DIRECTORY = f"{GIT_WORKING_DIRECTORY}/databags"
-
-
-def set_up_repo():
-    if not os.path.isdir(GIT_WORKING_DIRECTORY):
-        print("Cloning repo...")
-        repo = git.Repo.clone_from(
-            url=GITHUB_REPO_URL_WITH_CREDENTIALS, to_path=GIT_WORKING_DIRECTORY
-        )
-    else:
-        print("Setting up repo...")
-        repo = git.Repo(GIT_WORKING_DIRECTORY)
-    origin = repo.remotes.origin
-    print("Pulling latest code...")
-    origin.pull()
-    return f"{WORKING_DIRECTORY}/{GITHUB_REPO_NAME}"
-
-
-def set_up_git_user():
-    print("Setting up git user...")
-    os.environ["GIT_AUTHOR_NAME"] = GIT_USER_NAME
-    os.environ["GIT_AUTHOR_EMAIL"] = GIT_USER_EMAIL
-    os.environ["GIT_COMMITTER_NAME"] = GIT_USER_NAME
-    os.environ["GIT_COMMITTER_EMAIL"] = GIT_USER_EMAIL
 
 
 def set_up_github_client():
@@ -104,7 +63,7 @@ def get_repo_cc_metadata(repo):
     print("\tGetting CC metadata for this repo...")
     try:
         cc_metadata_file = repo.get_contents(CC_METADATA_FILE_NAME)
-    except UnknownObjectException:
+    except (UnknownObjectException, GithubException) as e:
         return {}
     cc_metadata = yaml.safe_load(cc_metadata_file.decoded_content)
     if "technologies" in cc_metadata:
@@ -145,48 +104,10 @@ def get_repo_data_dict(repo_data_list):
     return {"repos": repo_data_list}
 
 
-def generate_json_file(repo_data_dict):
-    print("Generating JSON file...")
-    json_filename = f"{JSON_FILE_DIRECTORY}/repos.json"
-    with open(json_filename, "w") as json_file:
-        json.dump(repo_data_dict, json_file, sort_keys=True, indent=4)
-    return json_filename
-
-
-def commit_and_push_changes(json_filename):
-    repo = git.Repo(GIT_WORKING_DIRECTORY)
-    git_diff = repo.index.diff(None)
-    if git_diff != []:
-        repo.index.add(items=f"{json_filename}")
-        repo.index.commit(message="Syncing new repository changes.")
-        origin = repo.remotes.origin
-        print("Pushing latest code...")
-        try:
-            origin.push()
-        except Exception as e:
-            print(f"Got exception {e} \n Trying manual push...")
-            g = git.Git(GIT_WORKING_DIRECTORY)
-            print(
-                g.execute(
-                    [
-                        "git",
-                        "push",
-                        f"{GITHUB_REPO_URL_WITH_CREDENTIALS}",
-                        "master",
-                    ]
-                )
-            )
-    else:
-        print("No changes to push...")
-
-
-if __name__ == "__main__":
-    set_up_repo()
-    set_up_git_user()
+def get_repo_data():
     github_client = set_up_github_client()
     cc = get_cc_organization(github_client)
     repos = get_repositories(cc)
     repo_data_list = get_repo_data_list(repos)
-    repo_data_dict = get_repo_data_dict(repo_data_list)
-    json_filename = generate_json_file(repo_data_dict)
-    commit_and_push_changes(json_filename)
+    data = get_repo_data_dict(repo_data_list)
+    return data
