--- search_roadmap_export/Pipfile
@@ -1,14 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-pygithub = "*"
-asana = "*"
-gitpython = "*"
-
-[requires]
-python_version = "3.7"
